Move orientation features to package.
