+++ scripts/data_exploration/player_orientation.py
@@ -5,6 +5,8 @@ import numpy as np
 import pandas as pd
 import seaborn as sns
 
+from mmfl.feature import orientation as orientation_feat
+
 # TODO create feature to map "continious" change without a flip between 180 and -180
 
 
@@ -23,41 +25,6 @@ def calc_start_orientation(play_df: pd.DataFrame, track_df: pd.DataFrame) -> pd.
     bool_result = possession_right | not_possession_left
 
     return pd.Series(np.where(bool_result, 90, 270), name="startOrientation")
-
-
-def _calc_left_start(orientation: pd.Series) -> np.ndarray:
-    # 270 deg
-    orientation_recalculated = np.zeros_like(orientation)
-    ind = orientation.between(90, 270)  # -180, 0
-    orientation_recalculated[ind] = orientation[ind] - 270
-    ind = orientation.between(270, 360)  # 0, 90
-    orientation_recalculated[ind] = orientation[ind] - 270
-    ind = orientation.between(0, 90)  # 90, 180
-    orientation_recalculated[ind] = orientation[ind] - 90
-    return orientation_recalculated
-
-
-def _calc_right_start(orientation: pd.Series) -> np.ndarray:
-    # 90 deg
-    orientation_recalculated = np.zeros_like(orientation)
-    ind = orientation.between(90, 270)  # 0, 180
-    orientation_recalculated[ind] = orientation[ind] - 90
-    ind = orientation.between(270, 360)  # -180, -90
-    orientation_recalculated[ind] = orientation[ind] - (180 + 270)
-    ind = orientation.between(0, 90)  # -90, 0
-    orientation_recalculated[ind] = orientation[ind] - 90
-    return orientation_recalculated
-
-
-def calc_orientation_deviation(
-    orientation_value: pd.Series, orientation_start: pd.Series
-) -> pd.Series:
-    orientation_recalculated = np.zeros_like(orientation_value)
-    is_right = orientation_start == 90
-    orientation_recalculated[is_right] = _calc_right_start(orientation_value[is_right])
-    is_left = orientation_start == 270
-    orientation_recalculated[is_left] = _calc_left_start(orientation_value[is_left])
-    return pd.Series(orientation_recalculated, name="orientationDeviation")
 
 
 def filter_od_line_players(
@@ -101,7 +68,7 @@ if __name__ == "__main__":
     tracking_df["startOrientation"] = calc_start_orientation(
         play_df=play_data_df, track_df=tracking_df
     )
-    tracking_df["orientationDeviation"] = calc_orientation_deviation(
+    tracking_df["orientationDeviation"] = orientation_feat.calc_orientation_deviation(
         tracking_df["o"], tracking_df["startOrientation"]
     )
     tracking_df["orientationDeviation"] = tracking_df["orientationDeviation"].abs()
