+++ src/mmfl/__init__.py
@@ -1,0 +1,3 @@
+from mmfl import feature
+
+__all__ = ["feature"]
