+++ src/mmfl/feature/__init__.py
@@ -1,0 +1,3 @@
+from mmfl.feature import orientation
+
+__all__ = ["orientation"]
